update ke sweet al;ert

+++ jscroot/js/controller/controller_logout.js
@@ -1,8 +1,20 @@
 import { deleteCookie } from 'https://jscroot.github.io/cookie/croot.js';
 
 function logout() {
-  deleteCookie('Login');
-  window.location.href = 'https://hrmonitor.advocata.me/dashboard/public/pages/login.html';
+  Swal.fire({
+    icon: 'warning',
+    title: 'Logout',
+    text: 'Are you sure you want to log out?',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, log out!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      deleteCookie('Login');
+      window.location.href = 'pages/login.html';
+    }
+  });
 }
 
 document.getElementById('logoutButton').addEventListener('click', logout);
